fix mistakes in SphericalTensor
