+++ tests/import_test.py
@@ -4,4 +4,5 @@
 def test_import():
     from e3nn.util import bounding_sphere, cache_file, default_dtype, time_logging  # noqa
     from e3nn.non_linearities import activation, gated_block, gated_block_parity, gru, norm_activation, norm, rescaled_act, s2, scalar_activation, so3  # noqa
-    from e3nn import batchnorm, dropout, groupnorm, kernel, linear, o3, radial, rs, spherical_tensor, rsh, tensor_product  # noqa
+    from e3nn import batchnorm, dropout, groupnorm, kernel, linear, o3, radial, rs, rsh, tensor_product  # noqa
+    from e3nn.tensor import irrep_tensor, spherical_tensor, fourrier_tensor  # noqa
