+++ tests/tensor/spherical_tensor_test.py
@@ -4,7 +4,8 @@ import unittest
 import torch
 
 from e3nn import o3, rs
-import e3nn.spherical_tensor as sphten
+from e3nn.tensor.spherical_tensor import spherical_harmonics_dirac, SphericalTensor, projection, adjusted_projection
+from e3nn.tensor.fourrier_tensor import FourierTensor
 
 
 class Tests(unittest.TestCase):
@@ -12,8 +13,8 @@ class Tests(unittest.TestCase):
         with o3.torch_default_dtype(torch.float64):
             for l in range(5):
                 r = torch.randn(3)
-                a = sphten.spherical_harmonics_dirac(r, l)
-                v = sphten.SphericalTensor(a, l).signal_xyz(r)
+                a = spherical_harmonics_dirac(r, l)
+                v = SphericalTensor(a, l).signal_xyz(r)
                 self.assertAlmostEqual(v.item(), 1)
 
     def test_projection(self):
@@ -21,14 +22,14 @@ class Tests(unittest.TestCase):
         lmax = 6
         coords = torch.randn(N, 3)
         coords = coords[coords.norm(2, -1) > 0]
-        sphten.projection(coords, lmax)
+        projection(coords, lmax)
 
     def test_adjusted_projection(self):
         N = 4
         lmax = 6
         coords = torch.randn(N, 3)
         coords = coords[coords.norm(2, -1) > 0]
-        sphten.adjusted_projection(coords, lmax)
+        adjusted_projection(coords, lmax)
 
     def test_plot_on_grid(self):
         pass
@@ -38,10 +39,9 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
     def test_SphericalTensor(self):
         torch.set_default_dtype(torch.float64)
         lmax = 6
-        mul = 1
-        sphten.SphericalTensor(torch.randn(mul * (lmax + 1) ** 2), lmax)
+        SphericalTensor(torch.randn((lmax + 1) ** 2), lmax)
         mul = 3
-        sphten.FourierTensor(torch.randn(mul * (lmax + 1) ** 2), mul, lmax)
+        FourierTensor(torch.randn(mul * (lmax + 1) ** 2), mul, lmax)
 
     def test_from_geometry(self):
         torch.set_default_dtype(torch.float64)
@@ -49,7 +49,7 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
         lmax = 6
         coords = torch.randn(N, 3)
         coords = coords[coords.norm(2, -1) > 0]
-        sphten.SphericalTensor.from_geometry(coords, lmax)
+        SphericalTensor.from_geometry(coords, lmax)
 
     def test_from_geometry_with_radial(self):
         torch.set_default_dtype(torch.float64)
@@ -61,18 +61,12 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
         def radial_model(x):
             return torch.ones_like(x).unsqueeze(-1)
 
-        sphten.FourierTensor.from_geometry(coords, radial_model, lmax)
+        FourierTensor.from_geometry(coords, radial_model, lmax)
 
     def test_sph_norm(self):
         torch.set_default_dtype(torch.float64)
         lmax = 6
-        mul = 1
-        sph = sphten.SphericalTensor(torch.randn(mul * (lmax + 1) ** 2), lmax)
-        sph.sph_norm()
-
-        mul = 3
-        sph = sphten.FourierTensor(torch.randn(
-            mul * (lmax + 1) ** 2), mul, lmax)
+        sph = SphericalTensor(torch.randn((lmax + 1) ** 2), lmax)
         sph.sph_norm()
 
     def test_plot(self):
@@ -81,7 +75,7 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
         lmax = 6
         coords = torch.randn(N, 3)
         coords = coords[coords.norm(2, -1) > 0]
-        sph = sphten.SphericalTensor.from_geometry(coords, lmax)
+        sph = SphericalTensor.from_geometry(coords, lmax)
 
         n = 16
         r, f = sph.plot(n=n)
@@ -98,9 +92,7 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
         def radial_model(x):
             return torch.ones_like(x).unsqueeze(-1)
 
-        sph = sphten.FourierTensor.from_geometry(coords,
-                                                 radial_model,
-                                                 lmax)
+        sph = FourierTensor.from_geometry(coords, radial_model, lmax)
 
         n = 16
         center = torch.ones(3)
@@ -111,33 +103,27 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
     def test_signal_on_sphere(self):
         torch.set_default_dtype(torch.float64)
         lmax = 4
-        sph = sphten.SphericalTensor(torch.randn((lmax + 1)**2), lmax)
+        sph = SphericalTensor(torch.randn((lmax + 1)**2), lmax)
 
         r, val1 = sph.signal_on_grid(2 * (lmax + 1))
         val2 = sph.signal_xyz(r)
         assert (val1 - val2).abs().max() < 1e-10
 
     def test_change_lmax(self):
-        lmax = 0
-        mul = 1
-        signal = torch.zeros(rs.dim([(mul, lmax)]))
-        sph = sphten.SphericalTensor(signal, lmax)
-        lmax_new = 5
-        sph_new = sph.change_lmax(lmax_new)
+        sph = SphericalTensor(torch.zeros(1), 0)
+        sph_new = sph.change_lmax(5)
         assert sph_new.signal.shape[0] == rs.dim(sph_new.Rs)
 
     def test_add(self):
         lmax = 4
-        mul = 1
         signal1 = torch.zeros((lmax + 1) ** 2)
         signal2 = signal1.clone()
         signal1[0] = 1.
         signal2[3] = 1.
-        sph1 = sphten.SphericalTensor(signal1, lmax)
-        sph2 = sphten.SphericalTensor(signal2, lmax)
+        sph1 = SphericalTensor(signal1, lmax)
+        sph2 = SphericalTensor(signal2, lmax)
 
         new_sph = sph1 + sph2
-        assert new_sph.mul == mul
         assert new_sph.lmax == max(sph1.lmax, sph2.lmax)
 
     def test_mul_and_dot(self):
@@ -146,8 +132,8 @@ class SphericalAndFourierTensorTests(unittest.TestCase):
         signal2 = signal1.clone()
         signal1[0] = 1.
         signal2[3] = 1.
-        sph1 = sphten.SphericalTensor(signal1, lmax)
-        sph2 = sphten.SphericalTensor(signal2, lmax)
+        sph1 = SphericalTensor(signal1, lmax)
+        sph2 = SphericalTensor(signal2, lmax)
 
         new_sph = sph1 * sph2
         assert rs.are_equal(new_sph.Rs, [(rs.mul_dim(sph1.Rs), 0, 0)])
